backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
drf-spectacular>=0.26.0
django-filter>=23.3
django-ratelimit>=4.1.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
passlib[argon2]>=1.7.4
cryptography>=41.0.0
httpx>=0.25.0
aiosmtplib>=3.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
python-dateutil>=2.8.0
pytz>=2023.3
factory-boy>=3.3.0
pytest-factoryboy>=2.6.1
django-prometheus==2.3.1
python-json-logger>=2.0.7
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-django==0.41b0
opentelemetry-exporter-jaeger==1.20.0
django-redis>=5.4.0
gunicorn>=21.2.0
requests>=2.31.0
